feat: ScheduleTableWrapper 컴포넌트 분리로 드롭시 리렌더링 개선
