+++ src/components/schedule-tables/__private__/index.ts
@@ -1,1 +1,2 @@
 export * from "./ScheduleTable";
+export * from "./ScheduleTableWrapper";
